aio: Remove iovecs from PgAioInProgress.
This is pretty nice, because due to combine support this was the
majority of struct PgAioInProgress size.

As it turns out, io_uring always copies the IOV during submit, making
it unnecessary to have a long live iovecs.

There's more to do to shrink PgAioInProgress to something sensible,
but not now...

Author:
Reviewed-By:
Discussion: https://postgr.es/m/
Backpatch: