aio: refactor test for whether you can scatter/gather.
Remove some more if-then-else spam from aio_io.c, when figuring out
whether an io_method can do scatter/gather.

While doing that, adjust it so that the answer can be different for
direct and buffered IO.  That's useful already for worker mode, which
might be able to give a more nuanced answer.  It's a hard requirement
for the future Windows mode, because Windows can't do scatter/gather
with buffered files.