aio: posix_aio: fix some races
CI was getting stuck because sometimes an IO would be reaped by not
complete, creating a deadling.  It's never OK to reap, not complete, and
go and do something else that might block.

XXX Need to tidy the code up to make that programming rule clearer.  In
the meantime, let's see if this gets CI to pass.