meson: ecpg: Split definition of static and shared libraries
Required for correct resource file generation, as the resource files should
only be added to the shared library.

This also fixes the copy-pasto of naming the target for the export files for
the ecpg libraries libpq.exports.

Only really makes sense with the subsequent commit, but is sufficiently
complicated to be worth keeping separate for easier review.