Issue advice for random streaming reads.
For random access to blocks that are not yet in our buffer pool, issue
POSIX_FADV_WILLNEED advice as soon as we have formed the largest range
we can.  (This is the point at which we'd start a real asynchronous
read, given more infrastructure.)

The total number of assumed-to-be-running I/Os is tracked and kept at or
below the requested number.

Even though we try to avoid generating purely sequential hints, based on
evidence that that is slower than just letting the kernel (at least on
Linux) detect sequential access, at least one future user needs a way to
disable advice, so provide a flag for that.