Sequential scans support streaming read
Add streaming read support for sequential scans. Do so by implementing
the streaming read callback to get the next block and save this block in
the scan descriptor.

The PgStreamingRead object is allocated in initscan(). This means it
will be freed and reallocated on rescan. Implementing a streaming read
reset function is a TODO for that API.

Currently, only forwards scans are supported by the streaming read API;
so, if a scan switches from forwards to backwards, the PgStreamingRead
object will need to be freed. This also means that if a scan switches
from backwards to forwards, it will not use streaming reads.

Distinguishing between a scan that has yet to be initialized, one that
doesn't support streaming reads and one that has switched scan
directions is one reason why it is difficult to wait until
heapfetchbuf() to allocate the PgStreamingRead object.