Sequential scans use Streaming Read API
Implementing streaming read support for sequential scans includes three
parts.

Allocate the streaming read object in heapgettup[_pagemode]() when it
has not yet been allocated, when it has been freed for a rescan, or when
the scan direction changes. Because heapgettup[_pagemode()] is in the
per tuple memory context, save the CurrentMemoryContext during scan
initialization and switch into it before allocating the streaming read
object.

Implement a callback returning the next block to prefetch to the
streaming read API.

Invoke the streaming read API when a new page is needed. Simplifying the
code to get the next page makes heapfetchbuf() short and simple enough
to inline.

ci-os-only: